nlp-service/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.4
pydantic-settings==2.3.4
spacy==3.7.5
PyMuPDF==1.24.5
python-docx==1.1.2
python-multipart==0.0.9
